+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 **Closed issues:**
 
-- Kamel install error [\#3348](https://github.com/apache/camel-k/issues/3348)
+- kamel local run support --build-property ? [\#3356](https://github.com/apache/camel-k/issues/3356)
 - E2E repeated failure [\#3315](https://github.com/apache/camel-k/issues/3315)
 - Integrations, IntegrationKits, IntegrationPlatforms are not owned by Namespaces [\#3311](https://github.com/apache/camel-k/issues/3311)
 - Helm Chart for camel-k 1.9.2 is invalid [\#3300](https://github.com/apache/camel-k/issues/3300)